Christmas birthdays are so rare
According to a 2006 Harvard University study, Christmas birthdays are so rare. December 25 is least common day for births (after February 29).
We shall see if there are fewer Nigerians (nairaland members) celebrating there birthdays on this date! But first lets use our web crawler to extract and save the dataset in a csv file.